pre-commit hook file for git repo
Users are requested to put this file in their .git/hooks directory with
the name pre-commit and with execute permissions enabled.  This will
enable automatic running of the genhostfabrics script on each checkin,
making sure that the host-fabrics file is always in sync with the defined
hosts.  Failure to keep this file up to date causes missing entries
in the /etc/hosts file and an inability to connect to a host's RDMA
devices by name.

Signed-off-by: Doug Ledford <dledford@redhat.com>