Initial import
This is not expected to work at this point.  I've mainly sanitized some
internal machine references and the like.  I have not done a thorough
audit for the types of changes needed to make this work in the FSDP
instead of internally in Red Hat's cluster.  Once we get the first
machines in the FSDP cluster, I'll fix this up while testing on those
machines.

Signed-off-by: Doug Ledford <dledford@redhat.com>